requirements.txt
attrs==22.2.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.2.2
python-dotenv==1.0.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
Werkzeug==2.2.2